requirements.txt
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
flask-marshmallow==0.14.0
marshmallow==3.14.1
Flask-JWT-Extended==3.24.1
PyJWT==1.7.1
Werkzeug==1.0.1
MarkupSafe==2.0.1
Jinja2==2.11.3
itsdangerous==1.1.0
click==7.1.2
SQLAlchemy>=1.4.48,<2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==0.19.2
pytest==8.3.5
